requirements.txt
numpy
pylops
matplotlib
scikit-image
color_matcher
pillow
scipy==1.12.0